# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Semillero",
        "description": "API para el sistema de Semillero",
        "contact": {
            "email": "tu_email@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/proyectos": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Consultar proyectos de un semillero",
                "description": "Devuelve todos los proyectos asociados a un semillero específico con información detallada",
                "operationId": "getProyectosPorSemillero",
                "parameters": [
                    {
                        "name": "semillero_id",
                        "in": "query",
                        "description": "ID del semillero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado del proyecto",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "inactivo",
                                "finalizado",
                                "en_revision"
                            ],
                            "example": "activo"
                        }
                    },
                    {
                        "name": "incluir_evaluaciones",
                        "in": "query",
                        "description": "Incluir evaluaciones de los proyectos",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de proyectos del semillero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "total_proyectos": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "semillero": {
                                                    "type": "string",
                                                    "example": "Semillero de IA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Semillero no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Semillero no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Registrar nuevo proyecto en un semillero (sin autenticación)",
                "description": "Registra un proyecto asociado a un semillero con todos los campos requeridos.",
                "operationId": "storeProyecto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "semillero_id",
                                    "descripcion",
                                    "lider_id",
                                    "coordinador_id",
                                    "estado",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Proyecto de Energía Solar"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Desarrollo de paneles solares para zonas rurales."
                                    },
                                    "semillero_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "lider_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "coordinador_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Activo"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-09"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "Proyecto de Energía Solar"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Desarrollo de paneles solares para zonas rurales."
                                        },
                                        "semillero_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "lider_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "coordinador_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "estado": {
                                            "type": "string",
                                            "example": "Activo"
                                        },
                                        "fecha_inicio": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-05-09"
                                        },
                                        "fecha_creacion": {
                                            "type": "string",
                                            "example": "2025-05-09T18:30:00.000000Z"
                                        },
                                        "fecha_actualizacion": {
                                            "type": "string",
                                            "example": "2025-05-09T18:30:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos proporcionados no son válidos."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/proyectos/{id}/asignar-estudiantes": {
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Asignar estudiantes a un proyecto",
                "description": "Permite asignar estudiantes a un proyecto específico.",
                "operationId": "1c7e04bc5805afb4376b8e1df3227b88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estudiantes"
                                ],
                                "properties": {
                                    "estudiantes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiantes asignados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiantes asignados correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proyecto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lista de estudiantes inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de estudiantes inválida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}